requirements.txt
discord.py>=2.3.2
anthropic>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
